exams now available at soecified time, session expiry added

+++ app.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 """flask app"""
+from flask import Flask, url_for, redirect, jsonify, flash, session
 import flask
-from flask import Flask, url_for, redirect, jsonify, flash
 from flask import render_template, request
 from models import storage
 from models.user import User
@@ -15,16 +15,13 @@ from flask_login import LoginManager, login_user, current_user, login_required, 
 from models.engine import query_functions
 from passlib.hash import bcrypt_sha256
 from dotenv import load_dotenv
-from datetime import datetime
+from datetime import datetime, timedelta
 import pytz
 import os
 
-
 load_dotenv()
-
 app = Flask(__name__)
 app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
-
 login_manager = LoginManager()
 login_manager.init_app(app)
 
@@ -89,7 +86,7 @@ def signin():
                 user.authenticated = True
                 storage.new(user)
                 storage.save()
-                login_user(user)
+                login_user(user, remember=False)
                 return redirect(url_for('dashboard'))
             else:
                 error = "Invalid email or password!"
@@ -114,7 +111,9 @@ def dashboard():
             return render_template("Instructor-dashboard.html")
         elif current_user.Role == 'Student':
             return redirect(url_for('student_dashboard'))
-    return flask.abort(401)
+
+    # return flask.abort(401)
+    return redirect(url_for('signin'))
 
 @login_required
 @app.route("/student-dashboard", methods=['GET'], strict_slashes=False)
@@ -127,7 +126,8 @@ def student_dashboard():
             available_exams.extend(exams)
         return render_template("Student-dashboard.html", exams=available_exams)
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/logout", strict_slashes=False)
@@ -139,12 +139,14 @@ def logout():
         storage.save()
         logout_user()
         return redirect(url_for('signin'))
-    return flask.abort(401)
+
+    return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/create_course", strict_slashes=False , methods=['GET', 'POST'])
 def create_course_form():
     if current_user.is_authenticated and current_user.Role == 'Instructor':
+        current_page = "Create Course"
         if request.method == 'POST':
             course_code = request.form['course_code']
             course_name = request.form['course_name']
@@ -156,32 +158,40 @@ def create_course_form():
             flash('Course successfully created', 'success')
             return redirect(url_for('view_courses'))
 
-        return render_template('create_course.html')
+        return render_template('create_course.html', current_page=current_page)
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/view-courses", methods=['GET'], strict_slashes=False)
 def view_courses():
     if current_user.is_authenticated and current_user.Role == 'Instructor':
+        current_page  = "View courses"
         courses = query_functions.get_user_courses(current_user.get_identification())
         exams = query_functions.get_all_exams_for_instructor(current_user.get_identification())
 
-        return render_template("view_course.html", courses=courses, exams=exams)
+        return render_template("view_course.html", courses=courses, exams=exams, current_page=current_page)
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @app.route("/dashboard/view-results", methods=["GET", "POST"])
 @login_required
 def view_students_results():
-    if request.method == "POST":
-        course_id = request.form['course_id']
-        exam_id = request.form["exam_id"]
-        results = query_functions.get_exam_results(exam_id)
-        return render_template("instructor-exam-results.html", results=results)
+    if current_user.is_authenticated and current_user.Role == 'Instructor':
+        current_page = "Exam performance"
+        if request.method == "POST":
+            course_id = request.form['course_id']
+            exam_id = request.form["exam_id"]
+            results = query_functions.get_exam_results(exam_id)
+            return render_template("instructor-exam-results.html", results=results, current_page=current_page)
+        else:
+            courses = query_functions.get_user_courses(current_user.get_identification())
+            return render_template("Instructor-results-form.html", courses=courses, current_page=current_page)
     else:
-        courses = query_functions.get_user_courses(current_user.get_identification())
-        return render_template("Instructor-results-form.html", courses=courses)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route('/get_exams_for_course')
@@ -196,51 +206,62 @@ def get_exams_for_course():
 @login_required
 @app.route("/dashboard/view-courses/delete_course/<int:course_id>", methods=['GET'], strict_slashes=False)
 def delete_course(course_id):
-    course_to_delete = query_functions.get_course(course_id)
-    if course_to_delete:
-        storage.delete(course_to_delete)
-        storage.save()
-        flash('Course successfully deleted', 'success')
+    if current_user.is_authenticated and current_user.Role == 'Instructor':
+        course_to_delete = query_functions.get_course(course_id)
+        if course_to_delete:
+            storage.delete(course_to_delete)
+            storage.save()
+            flash('Course successfully deleted', 'success')
+            return redirect(url_for('view_courses'))
+
         return redirect(url_for('view_courses'))
+    else:
+        return redirect(url_for('signin'))
 
-    return redirect(url_for('view_courses'))
-
-
+@login_required
 @app.route("/dashboard/view-course/view-questions/<int:exam_id>", methods=['GET', 'POST'], strict_slashes=False)
 def view_questions(exam_id):
-    questions = query_functions.get_questions(exam_id)
-    questionid = query_functions.get_exam(exam_id)
-    questionid = questionid.ExamID
-    return render_template("Created-questions.html", questions=questions, questionid=questionid)
+    if current_user.is_authenticated and current_user.Role == 'Instructor':
+        questions = query_functions.get_questions(exam_id)
+        questionid = query_functions.get_exam(exam_id)
+        questionid = questionid.ExamID
+        current_page = "View courses"
+        return render_template("Created-questions.html", questions=questions, questionid=questionid, current_page=current_page)
+    else:
+        return redirect(url_for('signin'))
 
-
+@login_required
 @app.route('/dashboard/view-courses/view-questions/delete_question/<int:exam_id>/<int:question_id>', methods=['POST'])
 def delete_question(question_id, exam_id):
-    # Fetch the question from the database
-    question = query_functions.get_question(question_id)
-    questions = query_functions.get_questions(exam_id)
-    if question:
-        storage.delete(question)
-        storage.save()
-        flash('Question deleted successfully', 'success')
-        return redirect(url_for('view_questions', questions=questions, exam_id=exam_id))
+    if current_user.is_authenticated and current_user.Role == 'Instructor':
+        question = query_functions.get_question(question_id)
+        questions = query_functions.get_questions(exam_id)
+        current_page = "View courses"
+        if question:
+            storage.delete(question)
+            storage.save()
+            flash('Question deleted successfully', 'success')
+            return redirect(url_for('view_questions', questions=questions, exam_id=exam_id, current_page=current_page))
+        else:
+            flash('Question not found', 'error')
+            return redirect(url_for('view_questions', questions=questions, current_page=current_page))
     else:
-        flash('Question not found', 'error')
-        return redirect(url_for('view_questions', questions=questions))
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/create_exam", methods=['GET', 'POST'], strict_slashes=False)
 @app.route("/dashboard/create_exam/<int:course_id>", methods=['GET', 'POST'], strict_slashes=False)
 def create_exam(course_id=None):
     if current_user.is_authenticated and current_user.Role == 'Instructor':
+        current_page = "Create Exam"
         if request.method == 'GET':
             available_courses = query_functions.get_user_courses(current_user.get_identification())
             course_exam = None
             if course_id:
                 course_exam = query_functions.get_course(course_id)
-                return render_template("create_exam.html", available_courses=available_courses, course_exam=course_exam)
+                return render_template("create_exam.html", available_courses=available_courses, course_exam=course_exam, current_page=current_page)
 
-            return render_template("create_exam.html", available_courses=available_courses, course_exam=course_exam)
+            return render_template("create_exam.html", available_courses=available_courses, course_exam=course_exam, current_page=current_page)
         elif request.method == 'POST':
             CourseID = request.form['selected-course-id']
             Title = request.form['exam_title']
@@ -256,7 +277,8 @@ def create_exam(course_id=None):
             flash('Exam successfully created', 'success')
             return redirect(url_for('view_courses'))
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/view-courses/delete_exam/<int:exam_id>", methods=['GET'], strict_slashes=False)
@@ -269,13 +291,15 @@ def delete_exam(exam_id):
             flash('Exam successfully deleted', 'success')
         return redirect(url_for('view_courses'))
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/questions-template/<int:exam_id>", methods=['GET', 'POST'])
 @app.route("/dashboard/questions-template", methods=['GET', 'POST'])
 def add_question(exam_id=None):
     if current_user.is_authenticated and current_user.Role == 'Instructor':
+        current_page = "Question Template"
         exams = query_functions.get_all_exams_for_instructor(current_user.get_identification())
         courses = query_functions.get_user_courses(current_user.get_identification())
         if request.method == 'POST':
@@ -308,12 +332,13 @@ def add_question(exam_id=None):
             storage.new(question)
             storage.save()
             flash('Question successfully created', 'success')
-            return redirect(url_for('add_question'))
+            return redirect(url_for('add_question', current_page=current_page))
 
         examination_id = exam_id
-        return render_template("Questions_template.html", exams=exams, courses=courses, examination_id=examination_id)
+        return render_template("Questions_template.html", exams=exams, courses=courses, examination_id=examination_id, current_page=current_page)
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/search_students_by_email")
@@ -329,12 +354,14 @@ def search_students_by_email():
         serialized = jsonify(serialized_students)
         return serialized
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/dashboard/register_students_template", methods=['GET', 'POST'])
 def register_students_template():
     if current_user.is_authenticated and current_user.Role == 'Instructor':
+        current_page = "Register Student"
         available_courses = query_functions.get_user_courses(current_user.get_identification())
         error = ""
         if request.method == 'POST':
@@ -345,17 +372,18 @@ def register_students_template():
             if existing_registration:
                 error =  "" #Student already registered to this course.
                 flash('Student already registered', 'error')
-                return render_template("Register_students.html", courses=available_courses)
+                return render_template("Register_students.html", courses=available_courses, current_page=current_page)
 
             studentcourse = StudentCourse(StudentID=StudentID, CourseID=CourseID)
             storage.new(studentcourse)
             storage.save()
             flash('Student successfully registered', 'success')
-            return redirect(url_for('register_students_template'))
+            return redirect(url_for('register_students_template', current_page=current_page))
         else:
-            return render_template("Register_students.html", courses=available_courses)
+            return render_template("Register_students.html", courses=available_courses, current_page=current_page)
     else:
-        flask.abort(401)
+        # flask.abort(401)
+        return redirect(url_for('signin'))
 
 @login_required
 @app.route("/take_exam/<int:exam_id>", methods=['GET', 'POST'])
@@ -368,15 +396,20 @@ def take_exam(exam_id):
             elif current_user.is_authenticated and current_user.Role == 'Student':
                 exam = query_functions.get_exam(exam_id)
                 if exam:
+                    current_time = datetime.utcnow()
                     exam_duration = exam.Duration
                     questions = exam.questions
                     total_questions = len(questions)
                     question_index = int(0)
-
-                    return render_template("Take-exam.html", questions=questions, question_index=question_index, exam_duration=exam_duration)
-                return render_template("Take-exam.html", questions=questions, question_index=question_index)
+                    if current_time >= exam.StartTime:
+                        return render_template("Take-exam.html", questions=questions, question_index=question_index, exam_duration=exam_duration)
+                    else:
+                        return "exam unavailable. check back later"
+                else:
+                    return render_template("Take-exam.html", questions=questions, question_index=question_index)
             else:
-                flask.abort(401)
+                # flask.abort(401)
+                return redirect(url_for('signin'))
     elif request.method == 'POST':
         if current_user.is_authenticated and current_user.Role == 'Student':
             exam = query_functions.get_exam(exam_id)
@@ -407,7 +440,7 @@ def take_exam(exam_id):
 
             return redirect(url_for('thank_you'))
         else:
-            flask.abort(401)
+            return redirect(url_for('signin'))
 
 @login_required
 @app.route("/student-dashboard/exam_results", strict_slashes=False)
@@ -415,6 +448,8 @@ def exam_results():
     if current_user.is_authenticated and current_user.Role == 'Student':
         user_results = query_functions.get_user_results(current_user.get_identification())
         return render_template("Results.html", results=user_results)
+    else:
+        return redirect(url_for('signin'))
 
 @app.route("/take_exam/submission", strict_slashes=False)
 def thank_you():
@@ -424,6 +459,17 @@ def thank_you():
 def add_header(response):
     response.cache_control.no_store = True
     return response
+
+@app.before_request
+def before_request():
+    session.permanent = True
+    app.permanent_session_lifetime = timedelta(minutes=1440)
+    session.modified = True
+
+@login_manager.unauthorized_handler
+@app.errorhandler(401)
+def unauthorized(error):
+    return redirect(url_for('signin'))
 
 if __name__ == "__main__":
     app.run(host='0.0.0.0', debug=True)
